+++ Documentation/lessons/3d_models/images/custom_models/0_set_scale.png
