+++ Python/example_controllers/3d_models/cube.fbx
